rename dialog component folder
