+++ src/app/management/routes/images/image-manager.component.ts
@@ -13,7 +13,7 @@ import { map } from 'rxjs/operators';
 import { Image } from 'src/app/models/entities/Image';
 import { ImageFormComponent } from 'src/app/shared/components/image-form/image-form.component';
 import { COMMON_DISMISS_BUTTON_LABEL, COMMON_ERROR_MESSAGE, COMMON_WARNING_MESSAGE } from 'src/text/messages';
-import { EntityFormDialogConfig } from '../../../shared/dialogs/entity-form-dialog/EntityFormDialogConfig';
+import { EntityFormDialogConfig } from '../../../shared/dialogs/entity-form/EntityFormDialogConfig';
 import { TransactionalDataManagerComponentDirective } from '../../directives/transactional-data-manager.component-directive';
 import { ImageManagerService } from './image-manager.service';
