+++ src/app/management/routes/shippers/shipper-manager.component.ts
@@ -13,7 +13,7 @@ import { map } from 'rxjs/operators';
 import { Shipper } from 'src/app/models/entities/Shipper';
 import { ShipperFormComponent } from 'src/app/shared/components/shipper-form/shipper-form.component';
 import { COMMON_WARNING_MESSAGE, COMMON_DISMISS_BUTTON_LABEL, COMMON_ERROR_MESSAGE } from 'src/text/messages';
-import { EntityFormDialogConfig } from '../../../shared/dialogs/entity-form-dialog/EntityFormDialogConfig';
+import { EntityFormDialogConfig } from '../../../shared/dialogs/entity-form/EntityFormDialogConfig';
 import { TransactionalDataManagerComponentDirective } from '../../directives/transactional-data-manager.component-directive';
 import { ShipperManagerService } from './shipper-manager.service';
