+++ src/app/shared/components/product-category-tree/product-category-tree.component.ts
@@ -16,8 +16,8 @@ import { ProductCategory } from 'src/app/models/entities/ProductCategory';
 import { COMMON_DISMISS_BUTTON_LABEL } from 'src/text/messages';
 import { ConfirmationDialogComponent } from '../../dialogs/confirmation/confirmation-dialog.component';
 import { ConfirmationDialogData } from '../../dialogs/confirmation/ConfirmationDialogData';
-import { EntityFormDialogComponent } from '../../dialogs/entity-form-dialog/entity-form-dialog.component';
-import { EntityFormDialogData } from '../../dialogs/entity-form-dialog/EntityFormDialogData';
+import { EntityFormDialogComponent } from '../../dialogs/entity-form/entity-form-dialog.component';
+import { EntityFormDialogData } from '../../dialogs/entity-form/EntityFormDialogData';
 import { ProductCategoryFormComponent } from '../product-category-form/product-category-form.component';
 import { ProductCategoryTreeService } from './product-category-tree.service';
 import { ProductCategoryTreeFlatNode } from './ProductCategoryTreeFlatNode';
