+++ src/app/shared/dialogs/entity-form/EntityFormDialogConfig.ts
