+++ src/app/shared/dialogs/entity-form/EntityFormDialogData.ts
